WIP on culling JVM stuff from core
