+++ app/src/main/kotlin/choliver/nespot/runner/InfoRunner.kt
@@ -1,6 +1,7 @@
 package choliver.nespot.runner
 
 import choliver.nespot.cartridge.Rom
+import choliver.nespot.printInfo
 
 
 class InfoRunner(private val rom: Rom) {
